+++ src/components/Close/Close.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import Close from '@components/Close'
+import { Close } from '@components/Close'
 
 # Close
