+++ src/components/Close/Close.test.tsx
@@ -1,6 +1,5 @@
-import Close from '@components/Close'
+import { Close } from '@components/Close'
 import { render } from '@testing-library/react'
-import React from 'react'
 
 const onClick = jest.fn()
