+++ src/components/Close/Close.tsx
@@ -1,0 +1,40 @@
+import Button from '@components/Button'
+import Icon from '@components/Icon'
+import { useStyles } from '@helpers/hooks/useStyles'
+import { FC } from 'react'
+
+import styles from './Close.module.scss'
+
+interface CloseProps {
+  className?: string | string[]
+  onClick?: () => void
+  iconName?: string
+  iconColor?: string
+}
+
+export const Close: FC<CloseProps> = ({
+  className = '',
+  onClick = () => null,
+  iconName = 'icon-remove',
+  iconColor
+}) => {
+  const elementStyles = useStyles(
+    {
+      [styles.close]: true
+    },
+    className
+  )
+
+  return (
+    <span className={elementStyles} onClick={onClick}>
+      <Button
+        variant='icon-transparent-hover'
+        data-testid='close-component-button'
+      >
+        <Icon icon={iconName} color={iconColor} />
+      </Button>
+    </span>
+  )
+}
+
+Close.displayName = 'Close'
